% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_logger.R
\name{logger_log_file}
\alias{logger_log_file}
\title{Return the current session log file path}
\usage{
logger_log_file()
}
\value{
Character path, or \code{NULL} if the logger has not been started.
}
\description{
Return the current session log file path
}
\seealso{
\code{\link[=start_session_logger]{start_session_logger()}}
}
