% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_logger.R
\name{logger_plot_dir}
\alias{logger_plot_dir}
\title{Return the current session plot directory}
\usage{
logger_plot_dir()
}
\value{
Character path, or \code{NULL} if the logger has not been started.
}
\description{
Return the current session plot directory
}
\seealso{
\code{\link[=start_session_logger]{start_session_logger()}}
}
