% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{Nour20}
\alias{Nour20}
\title{Nour20 full-spectrum color palette}
\format{
A named character vector of length 20.
}
\usage{
Nour20
}
\description{
A named vector of 20 colors sweeping the full visible spectrum from deep
navy through teal, green, gold, orange, red, and ending at indigo. Ideal
for single-cell datasets with many cell types (up to 20).
}
\keyword{datasets}
