% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{Nour_cols}
\alias{Nour_cols}
\title{Access colors from the Nour18 palette by name}
\usage{
Nour_cols(...)
}
\arguments{
\item{...}{Color names to retrieve. If empty, returns all 18 colors.}
}
\value{
A named character vector of hex color codes.
}
\description{
Access colors from the Nour18 palette by name
}
