% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{Nour_cols20}
\alias{Nour_cols20}
\title{Access colors from the Nour20 palette by name}
\usage{
Nour_cols20(...)
}
\arguments{
\item{...}{Color names to retrieve. If empty, returns all 20 colors.}
}
\value{
A named character vector of hex color codes.
}
\description{
Access colors from the Nour20 palette by name
}
