% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{Nour_pal}
\alias{Nour_pal}
\title{Return an interpolated scSidekick color palette function}
\usage{
Nour_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of palette in \code{Nour_palettes}. Default \code{"main"}.}

\item{reverse}{Logical. Reverse the palette direction. Default \code{FALSE}.}

\item{...}{Additional arguments passed to \code{\link[grDevices:colorRamp]{grDevices::colorRampPalette()}}.}
}
\value{
A palette function that takes an integer \code{n} and returns \code{n} colors.
}
\description{
Return an interpolated scSidekick color palette function
}
