% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{Nour_palettes}
\alias{Nour_palettes}
\title{Named list of scSidekick color palettes}
\format{
An object of class \code{list} of length 8.
}
\usage{
Nour_palettes
}
\description{
Palettes available:
\describe{
\item{main}{8 colors from Nour18: petrol, darkblue, darkpurple, darkpink, fuchsia, lightred, darkorange, lightorange}
\item{cool}{8 cool-toned colors from Nour18}
\item{hot}{8 warm-toned colors from Nour18}
\item{all}{All 18 colors from Nour18}
\item{spectrum}{All 20 colors from Nour20}
\item{blues}{5 blue-to-teal colors from Nour20}
\item{warm}{6 gold-to-crimson colors from Nour20}
\item{purples}{4 lavender-to-indigo colors from Nour20}
}
}
\keyword{datasets}
