% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_color_Nour}
\alias{scale_color_Nour}
\title{ggplot2 color scale using scSidekick palettes}
\usage{
scale_color_Nour(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of palette in \code{Nour_palettes}. Default \code{"main"}.}

\item{discrete}{Logical. Use a discrete scale? Default \code{TRUE}.}

\item{reverse}{Logical. Reverse the palette? Default \code{FALSE}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}} or
\code{\link[ggplot2:scale_gradient]{ggplot2::scale_color_gradientn()}}.}
}
\value{
A ggplot2 scale object.
}
\description{
ggplot2 color scale using scSidekick palettes
}
